Adds event helper to deal with firing of events
