+++ src/Library/Core.php
@@ -2,11 +2,7 @@
 
 namespace DrH\Tanda\Library;
 
-use DrH\Tanda\Events\TandaRequestEvent;
-use DrH\Tanda\Events\TandaRequestFailedEvent;
-use DrH\Tanda\Events\TandaRequestSuccessEvent;
 use DrH\Tanda\Exceptions\TandaException;
-use DrH\Tanda\Models\TandaRequest;
 use GuzzleHttp\Exception\GuzzleException;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Str;
@@ -153,19 +149,4 @@ class Core
         ];
     }
 
-    /**
-     * @param TandaRequest $request
-     * @return void
-     */
-    public function fireTandaEvent(TandaRequest $request): void
-    {
-        if ($request->status == 000001) {
-            event(new TandaRequestEvent($request));
-            return;
-        }
-
-        $request->status == 000000
-            ? TandaRequestSuccessEvent::dispatch($request)
-            : TandaRequestFailedEvent::dispatch($request);
-    }
 }
